requirements.txt
alembic==1.8.1
async-timeout==4.0.2
asyncio==3.4.3
bidict==0.22.0
blinker==1.4
bluedot==2.0.0
bluezero==0.7.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
chirpdat==1.0.2
chirpsdk==3.4.2
click==8.1.3
colorzero==2.0
configparser==3.5.0
cpufreq==0.3.5
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
dominate==2.7.0
entrypoints==0.3
eventlet==0.33.1
fail2ban==0.11.1
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-MQTT==1.1.1
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
getmac==0.8.3
ggwave==0.4.0
gpiozero==1.6.2
greenlet==2.0.1
grpcio==1.39.0
grpcio-tools==1.39.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
mutagen==1.45.1
netifaces==0.10.4
nmcli==1.1.1
numpy @ file://tmp/numpy-1.24.1.tar.gz
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paho-mqtt==1.6.1
pbr==5.9.0
Pillow==7.0.0
precise-lite-runner==0.4.0
precise-runner==0.3.1
protobuf==3.20.1
psutil==5.9.3
pvporcupine==2.1.3
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.6.1
pycups==1.9.73
pygame==1.9.6
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyserial==3.5
PySoundFile==0.9.0.post1
python-apt==2.0.0+ubuntu0.20.4.5
python-dotenv==0.19.2
python-engineio==4.3.4
python-gnupg==0.4.9
python-socketio==5.7.2
python-vlc==3.0.16120
pytz==2019.3
pyxattr==0.6.1
PyYAML==5.3.1
raspberrypi-tm1637==1.3.6
redis==4.3.5
requests==2.22.0
requests-futures==1.0.0
requests-unixsocket==0.2.0
RPi.GPIO==0.5.8
schedule==1.1.0
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sonopy==0.1.2
sounddevice==0.4.5
SpeechRecognition==3.8.1
speedtest-cli==2.1.3
SQLAlchemy==1.4.42
ssh-import-id==5.10
systemd-python==234
tflite-runtime==2.11.0
typing_extensions==4.3.0
ubuntu-advantage-tools==20.3
ufw==0.36
urllib3==1.25.8
visitor==0.1.3
wadllib==1.3.3
wavio==0.0.7
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.2
wget==3.2
wiringpi==2.60.1
youtube-dl==2020.3.24
zipp==3.8.0
